requirements.txt
flask==2.3.3
werkzeug==2.3.7
Pillow==9.5.0
gunicorn==20.1.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
gdown==4.7.1
scipy==1.10.1
